:white_check_mark: Add eventing tests
Ensure we check both instance and static members to ensure we add or
remove the event correctly for both cases.

Also added a tab completion test for event members, which we don't seem
to have explicit tests for that I could find. As this completion is
related to this PR, seemed appropriate to add tests for that behaviour
here as well.

Co-authored-by: Ilya <darpa@yandex.ru>